+++ src/main/java/io/fabric8/maven/docker/BuildMojo.java
@@ -1,6 +1,14 @@
 package io.fabric8.maven.docker;
 
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.LineNumberReader;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.net.URL;
 import java.util.Date;
+import java.util.Enumeration;
 
 import io.fabric8.maven.docker.access.DockerAccessException;
 import io.fabric8.maven.docker.config.BuildImageConfiguration;
@@ -23,6 +31,9 @@ import org.apache.maven.plugins.annotations.Parameter;
 @Mojo(name = "build", defaultPhase = LifecyclePhase.INSTALL)
 public class BuildMojo extends AbstractBuildSupportMojo {
 
+    public static final String DMP_PLUGIN_DESCRIPTOR = "META-INF/maven/io.fabric8/dmp-plugin";
+    public static final String DOCKER_EXTRA_DIR = "docker-extra";
+
     @Parameter(property = "docker.skip.build", defaultValue = "false")
     protected boolean skipBuild;
 
@@ -41,12 +52,14 @@ public class BuildMojo extends AbstractBuildSupportMojo {
             return;
         }
 
+        // Check for build plugins
+        executeBuildPlugins();
+
         // Iterate over all the ImageConfigurations and process one by one
         for (ImageConfiguration imageConfig : getResolvedImages()) {
             processImageConfig(hub, imageConfig);
         }
     }
-
 
     protected void buildAndTag(ServiceHub hub, ImageConfiguration imageConfig)
             throws MojoExecutionException, DockerAccessException {
@@ -64,11 +77,11 @@ public class BuildMojo extends AbstractBuildSupportMojo {
     }
 
     // We ignore an already existing date file and always return the current date
+
     @Override
     protected Date getReferenceDate() {
         return new Date();
     }
-
     private String determinePullPolicy(BuildImageConfiguration buildConfig) {
         return buildConfig != null && buildConfig.getImagePullPolicy() != null ? buildConfig.getImagePullPolicy() : imagePullPolicy;
     }
@@ -92,4 +105,61 @@ public class BuildMojo extends AbstractBuildSupportMojo {
             }
         }
     }
+
+    // check for a run-java.sh dependency an extract the script to target/ if found
+    private void executeBuildPlugins() {
+        try {
+            Enumeration<URL> dmpPlugins = Thread.currentThread().getContextClassLoader().getResources(DMP_PLUGIN_DESCRIPTOR);
+            while (dmpPlugins.hasMoreElements()) {
+
+                URL dmpPlugin = dmpPlugins.nextElement();
+                File outputDir = getAndEnsureOutputDirectory();
+                processDmpPluginDescription(dmpPlugin, outputDir);
+            }
+        } catch (IOException e) {
+            log.error("Cannot load dmp-plugins from %s", DMP_PLUGIN_DESCRIPTOR);
+        }
+    }
+
+    private void processDmpPluginDescription(URL pluginDesc, File outputDir) throws IOException {
+        String line = null;
+        try (LineNumberReader reader =
+                 new LineNumberReader(new InputStreamReader(pluginDesc.openStream(), "UTF8"))) {
+            line = reader.readLine();
+            while (line != null) {
+                if (line.matches("^\\s*#")) {
+                    // Skip comments
+                    continue;
+                }
+                callBuildPlugin(outputDir, line);
+                line = reader.readLine();
+            }
+        } catch (ClassNotFoundException e) {
+            // Not declared as dependency, so just ignoring ...
+        } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
+            log.verbose("Found dmp-plugin %s but could not be called : %s",
+                     line,
+                     e.getMessage());
+        }
+    }
+
+    private File getAndEnsureOutputDirectory() {
+        File outputDir = new File(new File(project.getBuild().getDirectory()), DOCKER_EXTRA_DIR);
+        if (!outputDir.exists()) {
+            outputDir.mkdirs();
+        }
+        return outputDir;
+    }
+
+    private void callBuildPlugin(File outputDir, String buildPluginClass) throws ClassNotFoundException, NoSuchMethodException, IllegalAccessException, InvocationTargetException {
+        Class buildPlugin = Class.forName(buildPluginClass);
+        try {
+            Method method = buildPlugin.getMethod("addExtraFiles", File.class);
+            method.invoke(null, outputDir);
+            log.info("Extra files from %s extracted", buildPluginClass);
+        } catch (NoSuchMethodException exp) {
+            log.verbose("Build plugin %s does not support 'addExtraFiles' method", buildPluginClass);
+        }
+    }
+
 }