+++ doc/changelog.md
@@ -2,13 +2,14 @@
 
 * **0.25-SNAPSHOT**
   - Always create missing target directory for docker:save (#1013)
+  - d-m-p plugins for adding extra files introduced
   - Update assembly plugin to 3.1.0 (#1021)
   - Add option for regenerating certificates after starting Docker Machine (#1019)
   - Add `startPeriod` to `healthCheck` (#961)
   - Unbreak setting of entrypoint in `exec` form when property mode is enabled (#1020)
   - Fix enabling of log configuration (#1010)
   - Add possiblity to use `docker.imagePropertyConfiguration` with multipe images (#1001)
-  
+
 * **0.25.2** (2018-04-14)
   - Fix for docker login issue with index.docker.io using a credential helper ([#946](https://github.com/fabric8io/docker-maven-plugin/issues/946))